+++ pkg/gatewayserver/config.go
@@ -39,7 +39,8 @@ type BasicStationConfig struct {
 
 // Config represents the Gateway Server configuration.
 type Config struct {
-	RequireRegisteredGateways bool `name:"require-registered-gateways" description:"Require the gateways to be registered in the Identity Server"`
+	RequireRegisteredGateways         bool          `name:"require-registered-gateways" description:"Require the gateways to be registered in the Identity Server"`
+	UpdateGatewayLocationDebounceTime time.Duration `name:"update-gateway-location-debounce-time" description:"Debounce time for gateway location updates from status messages"`
 
 	Stats                             GatewayConnectionStatsRegistry `name:"-"`
 	UpdateConnectionStatsDebounceTime time.Duration                  `name:"update-connection-stats-debounce-time" description:"Time before repeated refresh of the gateway connection stats"`